Trim the diff and correct inaccurate comments
Review follow-ups, no change to the PR's intent:

- Revert the ``predict_step`` hunk.  ``query_embed`` is a slice of ``last_hidden_state``,
  which ModernBERT's trailing LayerNorm keeps in fp32 under autocast, and the probs are
  already fp32 from ``logits_to_probs`` — so all three ``.float()`` calls were dead and the
  test guarding them passed with or without the change.  Drop both.
- Drop ``test_weights_stay_fp32_under_mixed_precision``: subsumed by the bf16-config test,
  and the only one of the pair that reaches the network.
- Fix the ``docformatter`` failure that would have turned Code Quality red.
- Correct three inaccurate claims: ``binary_cross_entropy_with_logits`` is on autocast's
  fp32 *cast* list, not the promote list; ``*-true`` are not the only precisions that cast
  weights (transformer-engine does too); and the matmul precision comment no longer implies
  it fixes a pre-existing divergence — both files were "medium" before, and this raises them
  together.
- Use the ``CONFIGS`` the module already exports, hoist test imports to module level, and
  pin the resume test to ``ALLOWED_DIFFERENCE_KEYS`` the way its sibling does.
- Shorten comments that ran well past the surrounding files' norm.

+++ src/every_query/model/lightning_module.py
@@ -404,19 +404,16 @@ class EveryQueryLightningModule(L.LightningModule):
         _, outputs = self.model(batch)
 
         empty_tensor = torch.tensor([])
-        # Probs come back fp32 already (``logits_to_probs`` upcasts before the sigmoid so bf16
-        # can't saturate them); ``query_embed`` is still bf16 under ``bf16-mixed`` and
-        # ``torch.Tensor.numpy()`` rejects bf16 outright, so keep the ``.float()`` on the way out.
         return {
-            "occurs_probs": outputs.occurs_probs.detach().float().cpu(),
-            "censor_probs": outputs.censor_probs.detach().float().cpu(),
+            "occurs_probs": outputs.occurs_probs.detach().cpu(),
+            "censor_probs": outputs.censor_probs.detach().cpu(),
             "occurs": batch.occurs.detach().cpu() if batch.occurs is not None else empty_tensor,
             "censor": batch.censor.detach().cpu() if batch.censor is not None else empty_tensor,
             "query": batch.query.detach().cpu() if batch.query is not None else empty_tensor,
             "duration_days": (
                 batch.duration_days.detach().cpu() if batch.duration_days is not None else empty_tensor
             ),
-            "query_embed": outputs.query_embed.detach().float().cpu(),
+            "query_embed": outputs.query_embed.detach().cpu(),
         }
 
     @staticmethod
